T #3 #4 Testing, error handling

+++ src/main/java/cz/amuradon/tvbot/WebhookData.java
@@ -1,11 +1,13 @@
 package cz.amuradon.tvbot;
 
+import java.math.BigDecimal;
+
 public record WebhookData(String symbol,
 		String side,
-		double quantity,
+		BigDecimal quantity,
 		boolean reduceOnly,
 		String newClientOrderId,
-		double stopLoss,
+		BigDecimal stopLoss,
 		String userUuid) {
 
 }
